requirements.txt
streamlit
pandas
folium
streamlit-folium
scikit-learn
google-generativeai